feat: added customer search and service overview retrieval

+++ src/main/java/com/booker_app/backend_service/controllers/request/ServiceRequest.java
@@ -2,6 +2,8 @@
 Booker App. */
 package com.booker_app.backend_service.controllers.request;
 
+import java.math.BigDecimal;
+
 import com.booker_app.backend_service.models.ServiceLength;
 import lombok.*;
 
@@ -12,7 +14,7 @@ import lombok.*;
 public class ServiceRequest {
 	@NonNull private String name;
 	private String description;
-	@NonNull private Double price;
+	@NonNull private BigDecimal price;
 	@NonNull private Integer time;
 	@NonNull private ServiceLength serviceLength;
 }
